:art: Move fixtures only needed for integration tests under the `integration` directory
